requirements.txt
aiogram>=3.4
aiosqlite>=0.19
apscheduler>=3.10
aiohttp>=3.9
python-dotenv>=1.0
python-dateutil>=2.8
